Now uses floats instead of doubles for exported data. Made it eeasy to switch via a typedef.

+++ src/bufr/BufrParser/BufrCollectors/BufrAccumulator.cpp
@@ -21,7 +21,7 @@ namespace Ingester
     {
     }
 
-    void BufrAccumulator::addRow(std::vector<double>& newRow)
+    void BufrAccumulator::addRow(std::vector<FloatType>& newRow)
     {
         if (numDataRows_ + 1 > dataArray_.rows())
         {
